feat: 프로필 이미지 추가 기능 구현

+++ server/src/main/java/com/codestates/server/domain/member/service/MemberService.java
@@ -1,7 +1,9 @@
 package com.codestates.server.domain.member.service;
 
+import com.codestates.server.domain.member.entity.AuthUserUtils;
 import com.codestates.server.domain.member.entity.Member;
 import com.codestates.server.domain.member.repository.MemberRepository;
+import com.codestates.server.domain.member.s3.S3UploadService;
 import com.codestates.server.global.exception.BusinessLogicException;
 import com.codestates.server.global.exception.ExceptionCode;
 import com.codestates.server.global.security.auth.utils.CustomAuthorityUtils;
@@ -11,7 +13,9 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -26,6 +30,11 @@ public class MemberService {
     private final PasswordEncoder passwordEncoder;
     // 사용자 권한 설정
     private final CustomAuthorityUtils customAuthorityUtils;
+    private final S3UploadService s3UploadService;
+    private static final String DEFAULT_IMAGE = "http://bit.ly/46a2mSp";
+    private static final String MEMBER_IMAGE_PROCESS_TYPE = "profile-image";
+
+
 
     /**
      * 회원 가입 로직
@@ -47,6 +56,12 @@ public class MemberService {
         List<String> roles = customAuthorityUtils.createRoles(member.getEmail());
         member.setRoles(roles);
 
+        if(member.getProfileImage() == null) {
+            member.setProfileImage(DEFAULT_IMAGE);
+        } else {
+            member.setProfileImage(member.getProfileImage());
+        }
+
         // 예외 발생 안 시키면 저장
         Member savedMember = memberRepository.save(member);
 
@@ -58,6 +73,10 @@ public class MemberService {
         // 없는 회원이면 예외 발생
         Member getMember = getVerifiedMember(member.getMemberId());
 
+        // 로그인한 MemberId랑 업데이트 하려는 마이페이지의 memberId랑 다르면 예외 발생
+        if(!getLoginMember().getMemberId().equals(getMember.getMemberId()))
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);
+
         Optional.ofNullable(member.getName())
                 .ifPresent(nickname -> getMember.setName(member.getName()));
         Optional.ofNullable(member.getPassword())
@@ -66,6 +85,52 @@ public class MemberService {
                 .ifPresent(image -> getMember.setProfileImage(member.getProfileImage()));
 
         return memberRepository.save(getMember);
+    }
+
+//    public String uploadImage(Long memberId, MultipartFile file) {
+    public String uploadImage(Long memberId, MultipartFile file, int x, int y, int width, int height) throws IOException {
+        // 회원 검증 하기
+        Member member = getVerifiedMember(memberId);
+        // 현재 프로필 이미지 가지고 오기
+        String presentProfileImage = member.getProfileImage();
+        // 만약에 현재 파일이 있으면 현재 파일 삭제
+        if (presentProfileImage != null && !presentProfileImage.equals(DEFAULT_IMAGE)) {
+//            s3UploadService.deleteImageFromS3(presentProfileImage, MEMBER_IMAGE_PROCESS_TYPE);
+            s3UploadService.deleteImageFromS3(presentProfileImage);
+        }
+        // profileImage 새로운 imgae로 upload 하기
+        String newProfileImage = null;
+        // 새로운 파일 업로드 하는 메서드 (파일, x좌표, y좌표, 가로, 세로)
+//        newProfileImage = s3UploadService.uploadProfileImage(file, MEMBER_IMAGE_PROCESS_TYPE);
+        newProfileImage = s3UploadService.uploadProfileImage(file, x, y, width, height);
+        // 회원 profileImage에 set 하고 save
+        member.setProfileImage(newProfileImage);
+        // 새로운 이미지 URL을 멤버 객체에 설정
+        member.setProfileImage(newProfileImage);
+        // 회원 정보 업데이트
+        memberRepository.save(member);
+
+        return newProfileImage;
+    }
+
+    public String deleteProfileImage(Long memberId) {
+        // 회원 검증 하기
+        Member member = getVerifiedMember(memberId);
+
+        // 현재 프로필 이미지 URL을 가져옵니다.
+        String currentProfileImage = member.getProfileImage();
+
+        // S3에서 현재 이미지 삭제
+        if (currentProfileImage != null && !currentProfileImage.equals(DEFAULT_IMAGE)) {
+//            s3UploadService.deleteImageFromS3(currentProfileImage, MEMBER_IMAGE_PROCESS_TYPE);
+            s3UploadService.deleteImageFromS3(currentProfileImage);
+        }
+
+        // DB에서 프로필 이미지 URL 기본 설정
+        member.setProfileImage(DEFAULT_IMAGE);
+        memberRepository.save(member);
+
+        return DEFAULT_IMAGE;
     }
 
     // member 사용자 정보 가지고 오는 메서드
@@ -85,6 +150,10 @@ public class MemberService {
     public void deleteMember(Long memberId) {
         Member getMember = getVerifiedMember(memberId);
 
+        // 만약 로그인 한 사용자랑 회원 삭제 하려는 회원이랑 memberId 다르면 예외 던지기
+        if(!getLoginMember().getMemberId().equals(getMember.getMemberId()))
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);
+
         memberRepository.delete(getMember);
     }
 
@@ -99,7 +168,8 @@ public class MemberService {
         Optional<Member> member = memberRepository.findById(memberId);
 
         // 회원이 아니면 예외 발생
-        Member getMember = member.orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
+        Member getMember =
+                member.orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
 
         return getMember;
     }
@@ -114,8 +184,18 @@ public class MemberService {
 
         Optional<Member> optionalMember = memberRepository.findByEmail(email);
 
-        if(optionalMember.isPresent()) throw new BusinessLogicException(ExceptionCode.USER_EXISTS);
+        if(optionalMember.isPresent())
+            throw new BusinessLogicException(ExceptionCode.USER_EXISTS);
 
+    }
+
+    /**
+     * 로그인한 Member를 가지고 오는 메서드
+     * @return
+     */
+    public Member getLoginMember() {
+        return memberRepository.findByEmail(AuthUserUtils.getAuthUser().getName())
+                .orElseThrow(()-> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
     }
 
 }
